Resolved issue #1 - Preview page not updating on re-edit
When a new, draft, or existing post is re-edit after the first edit and
previewurl generation, each successive time the previewurl is shown, it
will be the same URL. This was occuring before the fix, but the file at
the previewurl was never getting updated (accept for the firs time at
previewurl generation). This now works.

Required updates to both functions do_edit and make_preview.